edit and save place

+++ src/components/others/Table.js
@@ -53,17 +53,20 @@ export default function CustomizedTables(props) {
   const classes = useStyles();
   const dispatch = useDispatch();
 
-  const {places, handle, setOpenModal, setLocation, setTitle, setId} = props
+  const {places, handle, setOpenModal, setLocation, setId} = props
 
   const handleOpenModal = (place) => {
-    setTitle(place.title)
     setLocation(place)
     setId(place.placeId)
     setOpenModal(true)
   }
 
   const handleDelete = (placeId) => {
-    dispatch(deletePlace(placeId))
+    if (window.confirm('Tem certeza que deseja deletar esta localização')) {
+      dispatch(deletePlace(placeId))
+    } else {
+      return false
+    }
   }
 
   return (
